feat: add support to ping on connections though proxy
Add a capability to test if the connection was established on the
other side of the proxy/tunnel by sending and/or receiving data through
the socket.

The server response latency will influence the final result.

Added the following arguments :
--proxy: Enable the test.

--proxy-write-buffer: Allow you to set a buffer to be written on the
socket. If not provided, the test nothing will be written.

--proxy-read-len: The size of the buffer to be read from the socket. If
zero is passed, the read test will be skipped.